refactor: move search box into another component
